+++ meta/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+# [36.33] - 2026-2-27
+-  added system states v2
+-  added economic weather interface
+
 # [36.31] - 2026-2-26
 -  added map screen POI indicators